Fix bugs

+++ src/models/project.model.js
@@ -3,6 +3,7 @@ import { ProjectMember } from "./projectmember.model.js";
 import { ApiError } from "../utils/api-error.js";
 import { ProjectNote } from "./note.model.js";
 import { Task } from "./task.model.js";
+import { SubTask } from "./subtask.model.js";
 
 const projectSchema = new Schema(
   {
@@ -24,14 +25,15 @@ const projectSchema = new Schema(
   { timestamps: true },
 );
 
-projectSchema.post("findOneAndDelete", async (project, next) => {
-  try {
-    await ProjectMember.deleteMany({ project: project._id });
-    await ProjectNote.deleteMany({ project: project._id });
-    await Task.deleteMany({ project: project._id });
-    next();
-  } catch (error) {
-    throw new ApiError(400, "Some Error occurred");
+projectSchema.post("findOneAndDelete", async (project) => {
+  if (project) {
+    try {
+      await ProjectMember.deleteMany({ project: project._id });
+      await ProjectNote.deleteMany({ project: project._id });
+      await Task.deleteMany({ project: project._id });
+    } catch (error) {
+      throw new ApiError(400, "Some Error occurred");
+    }
   }
 });
